datacube: separate datacube from REPL code
