+++ packages/legend-application-repl/src/stores/LegendREPLDataCubeApplicationEngine.ts
@@ -19,8 +19,8 @@ import {
   APPLICATION_EVENT,
 } from '@finos/legend-application';
 import { type DocumentationEntry, LogEvent } from '@finos/legend-shared';
-import { DataCubeApplicationEngine } from './dataCube/DataCubeApplicationEngine.js';
-import type { ActionAlert } from '../components/shared/Alert.js';
+import { DataCubeApplicationEngine } from './dataCube/engine/DataCubeApplicationEngine.js';
+import type { ActionAlert } from '../components/dataCube/application/DataCubeAlert.js';
 import type { LegendREPLApplicationStore } from '../application/LegendREPLApplicationStore.js';
 
 export class LegendREPLDataCubeApplicationEngine extends DataCubeApplicationEngine {
