+++ packages/legend-application-repl/src/stores/dataCube/core/DataCubeQuerySnapshotManager.ts
@@ -22,7 +22,7 @@ import {
   deepDiff,
   guaranteeNonNullable,
 } from '@finos/legend-shared';
-import type { DataCubeQuery } from '../../../server/DataCubeQuery.js';
+import type { DataCubeQuery } from '../engine/DataCubeQuery.js';
 
 // TODO: set a stack depth when we implement undo/redo
 // const DATA_CUBE_MAX_SNAPSHOT_COUNT = 100;
